refactor(form-v2): update links
